[FIX] 크래프팅 리스트 창 업데이트 안되던 버그 수정

+++ Source/AWeek/Components/AWeekInventoryComponent.cpp
@@ -466,7 +466,7 @@ int32 UAWeekInventoryComponent::FindFirstEmptySlot()
 	return -1;
 }
 
-void UAWeekInventoryComponent::UpdateInventoryTotalWeight(int32 DeltaWeight)
+void UAWeekInventoryComponent::UpdateInventoryTotalWeight(float DeltaWeight)
 {
 	bool bIsEncumbered = InventoryTotalWeight > InventoryWeightCapacity;
 	InventoryTotalWeight += DeltaWeight;
